Added ECPPartitioner-Draft Corrected top-k match

+++ evaluation/src/main/scala/ch/unibas/dmi/dbis/adam/evaluation/EvaluationResultLogger.scala
@@ -17,7 +17,7 @@ trait EvaluationResultLogger {
   val out = new PrintWriter(bw)
 
   def appendToResults(tuples: Int, dimensions: Int, partitions: Int, index: String, time: Float, k: Int = 0, noResults: Int = 0, partitioner: RepartitionMessage.Partitioner, informationLoss: Double, dropPerc: Double, ratio: Float): Unit = {
-    out.println(Calendar.getInstance().getTime() + "," + index + "," + tuples + "," + dimensions + "," + partitions + "," + time + "," + k + "," + noResults + "," + partitioner.name+","+dropPerc+","+informationLoss.toInt+","+ratio)
+    out.println(Calendar.getInstance().getTime() + "," + index + "," + tuples + "," + dimensions + "," + partitions + "," + time + "," + k + "," + noResults + "," + partitioner.name+","+dropPerc+","+informationLoss+","+ratio)
     out.flush()
   }
 }
